requirements.txt
# MLB AI Analytics Platform
# Install: pip install -r requirements.txt

# =============================================================================
# Deep Learning - PyTorch
# =============================================================================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# =============================================================================
# Machine Learning
# =============================================================================
scikit-learn>=1.4.0
xgboost>=2.0.0
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.12.0

# =============================================================================
# MLOps
# =============================================================================
mlflow>=2.9.0
optuna>=3.5.0              # Hyperparameter optimization
dvc>=3.30.0                # Data version control

# =============================================================================
# LLM / NLP
# =============================================================================
openai>=1.10.0
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.0            # Vector database for RAG
tiktoken>=0.5.0
transformers>=4.36.0
peft>=0.7.0                # Parameter-efficient fine-tuning (LoRA)
accelerate>=0.25.0
bitsandbytes>=0.42.0       # 8-bit/4-bit quantization

# =============================================================================
# Cloud - AWS
# =============================================================================
boto3>=1.34.0
sagemaker>=2.200.0

# =============================================================================
# Cloud - GCP
# =============================================================================
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.14.0
kfp>=2.5.0                 # Kubeflow Pipelines

# =============================================================================
# Web & API
# =============================================================================
streamlit>=1.29.0
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0

# =============================================================================
# Data Sources
# =============================================================================
MLB-StatsAPI>=1.7.2
pybaseball>=2.2.7

# =============================================================================
# Visualization
# =============================================================================
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.7

# =============================================================================
# Testing
# =============================================================================
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# =============================================================================
# Code Quality
# =============================================================================
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.1.9
